Clean site and modularize (:wip)
