+++ serialization_jackson_json/build.gradle.kts
@@ -10,8 +10,6 @@ apply(from = "../gradle/detekt.gradle")
 
 description = "Hexagon JSON serialization format (using Jackson)."
 
-extra["basePackage"] = "com.hexagonkt.serialization.jackson.json"
-
 dependencies {
     api(project(":serialization_jackson"))
     testImplementation(project(":serialization_test"))
