+++ site/pages/examples/templates_examples.md
@@ -1,16 +1,16 @@
 
 # Register Template Adapters
 Example to show how to register template adapters.
-[full test](https://github.com/hexagonkt/hexagon/blob/master/templates/src/test/kotlin/examples/TemplatesTest.kt).
+[full test](https://github.com/hexagonkt/hexagon/blob/master/templates/src/test/kotlin/com/hexagonkt/templates/examples/TemplatesTest.kt).
 
 ## Using regular Expressions
-@code templates/src/test/kotlin/examples/TemplatesTest.kt?templateRegex
+@code templates/src/test/kotlin/com/hexagonkt/templates/examples/TemplatesTest.kt?templateRegex
 
 ## Using globs
-@code templates/src/test/kotlin/examples/TemplatesTest.kt?templateGlob
+@code templates/src/test/kotlin/com/hexagonkt/templates/examples/TemplatesTest.kt?templateGlob
 
 # Process Templates
 Simple example to show how to process templates.
-[full test](https://github.com/hexagonkt/hexagon/blob/master/templates/src/test/kotlin/examples/TemplatesTest.kt).
+[full test](https://github.com/hexagonkt/hexagon/blob/master/templates/src/test/kotlin/com/hexagonkt/templates/examples/TemplatesTest.kt).
 
-@code templates/src/test/kotlin/examples/TemplatesTest.kt?templateUsage
+@code templates/src/test/kotlin/com/hexagonkt/templates/examples/TemplatesTest.kt?templateUsage
