+++ starters/build.gradle.kts
@@ -50,7 +50,6 @@ extensions.configure<PublishingExtension> {
     }
 }
 
-// TODO Move these methods to `buildSrc` and consider using dom4j or jdom
 fun File.parseDom(): Document =
     DocumentBuilderFactory.newInstance().newDocumentBuilder().parse(this)
