+++ templates/src/main/kotlin/com/hexagonkt/templates/TemplateManager.kt
