+++ templates/src/test/kotlin/com/hexagonkt/templates/SampleTemplateAdapter.kt
