+++ templates/src/test/kotlin/com/hexagonkt/templates/TemplateManagerTest.kt
