+++ templates/src/test/kotlin/com/hexagonkt/templates/examples/TemplatesTest.kt
