+++ templates_freemarker/build.gradle.kts
@@ -7,8 +7,6 @@ apply(from = "../gradle/kotlin.gradle")
 apply(from = "../gradle/publish.gradle")
 apply(from = "../gradle/dokka.gradle")
 
-extra["basePackage"] = "com.hexagonkt.templates.freemarker"
-
 dependencies {
     api(project(":templates"))
     api("org.freemarker:freemarker:${properties["freemarkerVersion"]}")
