+++ templates_freemarker/src/main/kotlin/com/hexagonkt/templates/freemarker/FreeMarkerAdapter.kt
