+++ templates_freemarker/src/test/kotlin/com/hexagonkt/templates/freemarker/FreeMarkerAdapterTest.kt
