+++ templates_test/build.gradle.kts
@@ -7,8 +7,6 @@ apply(from = "../gradle/kotlin.gradle")
 apply(from = "../gradle/publish.gradle")
 apply(from = "../gradle/dokka.gradle")
 
-extra["basePackage"] = "com.hexagonkt.templates"
-
 dependencies {
     val kotlinVersion = properties["kotlinVersion"]
     val junitVersion = properties["junitVersion"]
